requirements.txt
openai
google-generativeai
anthropic
